+++ packages/office-ui-fabric-react/src/components/Button/BaseButton.tsx
@@ -15,7 +15,7 @@ import { DirectionalHint } from '../../common/DirectionalHint';
 import { ContextualMenu, IContextualMenuProps } from '../../ContextualMenu';
 import { IButtonProps, IButtonClassNames, IButton } from './Button.Props';
 import styles = require('./BaseButton.scss');
-import { customizable } from '../ThemeProvider/customizable';
+import { customizable } from '@uifabric/utilities/lib/customizable';
 
 export interface IBaseButtonState {
   menuProps?: IContextualMenuProps | null;
