+++ packages/office-ui-fabric-react/src/components/Button/IconButton/IconButton.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { BaseButton } from '../BaseButton';
 import { BaseComponent, IBaseProps } from '@uifabric/utilities';
 import { IButtonProps, IButtonClassNames } from '../Button.Props';
-import { customizable } from '../../ThemeProvider/customizable';
+import { customizable } from '@uifabric/utilities/lib/customizable';
 import styles = require('./IconButton.scss');
 
 export const IconButtonClassNames = {
